# openapi.yaml
...
openapi: 3.1.0
info:
  title: photolib
  description: "\nA local, private photo library. Natural-language search over your photos,\nface search\
    \ and recognition, and date/location browsing \u2014 all running on your\nown machine with no network\
    \ calls.\n"
  version: 2.0.0
paths:
...
  /api/v1/admin/roots:
    get:
      tags:
      - admin
      summary: List Roots
      description: Every folder the user has added to the library, with photo counts.
      operationId: list_roots_api_v1_admin_roots_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - admin
      summary: Add Root
      description: Remember a source folder (indexing it is a separate, explicit step).
      operationId: add_root_api_v1_admin_roots_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RootRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Remove Root
      description: Forget a source folder. Photos already indexed from it stay.
      operationId: remove_root_api_v1_admin_roots_delete
      parameters:
      - name: path
        in: query
        required: true
        schema:
          type: string
          description: Folder to forget
          title: Path
        description: Folder to forget
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/index:
    post:
      tags:
      - admin
      summary: Start Index
      description: 'Index (or re-index) a folder in the background.


        Returns immediately with a job id; poll ``/admin/jobs/{id}`` for progress.

        Indexing is incremental, so pointing this at the same folder after adding

        photos only costs model time for the new files.'
      operationId: start_index_api_v1_admin_index_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/recluster:
    post:
      tags:
      - admin
      summary: Start Recluster
      description: 'Rebuild every person from scratch from the stored face embeddings.


        Useful after tuning thresholds. Manually confirmed faces are preserved.'
      operationId: start_recluster_api_v1_admin_recluster_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReclusterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/jobs/{job_id}:
    get:
      tags:
      - admin
      summary: Get Job
      operationId: get_job_api_v1_admin_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Cancel Job
      operationId: cancel_job_api_v1_admin_jobs__job_id__delete
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/curation:
    get:
      tags:
      - admin
      summary: Export Curation
      description: 'Create a checksummed backup of all hand-made curation.


        Photos are matched by content hash and path. Confirmed face assignments

        also carry their exact bounding boxes, preserving identity splits and

        confirmed unassigned faces across a rebuild when those faces still match.'
      operationId: export_curation_api_v1_admin_curation_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - admin
      summary: Import Curation
      description: Verify and restore matched curation without guessing conflicts.
      operationId: import_curation_api_v1_admin_curation_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Data
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/duplicates:
    get:
      tags:
      - admin
      summary: Duplicates
      description: Groups of identical or near-identical photos, for reclaiming space.
      operationId: duplicates_api_v1_admin_duplicates_get
      parameters:
      - name: max_distance
        in: query
        required: false
        schema:
          type: integer
          maximum: 20
          minimum: 0
          default: 6
          title: Max Distance
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/annotation:
    patch:
      tags:
      - curation
      summary: Annotate
      operationId: annotate_api_v1_images__image_id__annotation_patch
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Annotation'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/images/{image_id}/burst:
    get:
      tags:
      - curation
      summary: Burst
      operationId: burst_api_v1_images__image_id__burst_get
      parameters:
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/saved-searches:
    get:
      tags:
      - curation
      summary: Searches
      operationId: searches_api_v1_saved_searches_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - curation
      summary: Save Search
      operationId: save_search_api_v1_saved_searches_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSearch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/saved-searches/{search_id}:
    delete:
      tags:
      - curation
      summary: Delete Search
      operationId: delete_search_api_v1_saved_searches__search_id__delete
      parameters:
      - name: search_id
        in: path
        required: true
        schema:
          type: string
          title: Search Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/roots/relocate:
    post:
      tags:
      - curation
      summary: Relocate
      operationId: relocate_api_v1_admin_roots_relocate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Relocate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/curation/verify:
    post:
      tags:
      - curation
      summary: Verify Backup
      operationId: verify_backup_api_v1_admin_curation_verify_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Data
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Annotation:
      properties:
        favorite:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Favorite
        rating:
          anyOf:
          - type: integer
            maximum: 5.0
            minimum: 0.0
          - type: 'null'
          title: Rating
      type: object
      title: Annotation
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IndexRequest:
      properties:
        folder:
          type: string
          title: Folder
          description: Absolute path to a folder of photos
        rebuild:
          type: boolean
          title: Rebuild
          description: Reprocess this folder, preserving IDs and curation; models must remain compatible
          default: false
        prune_missing:
          type: boolean
          title: Prune Missing
          description: Remove indexed photos whose files are gone
          default: false
      type: object
      required:
      - folder
      title: IndexRequest
    JobOut:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          title: Kind
        status:
          type: string
          title: Status
        phase:
          type: string
          title: Phase
          default: ''
        current:
          type: integer
          title: Current
          default: 0
        total:
          type: integer
          title: Total
          default: 0
        percent:
          type: number
          title: Percent
          default: 0.0
        elapsed:
          type: number
          title: Elapsed
          default: 0.0
        detail:
          additionalProperties: true
          type: object
          title: Detail
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - id
      - kind
      - status
      title: JobOut
...
    ReclusterRequest:
      properties:
        threshold:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Threshold
        knn:
          anyOf:
          - type: integer
            maximum: 200.0
            minimum: 2.0
          - type: 'null'
          title: Knn
      type: object
      title: ReclusterRequest
    Relocate:
      properties:
        old_folder:
          type: string
          title: Old Folder
        new_folder:
          type: string
          title: New Folder
        verification:
          anyOf:
          - type: string
          - type: 'null'
          title: Verification
      type: object
      required:
      - old_folder
      - new_folder
      title: Relocate
...
    RootRequest:
      properties:
        folder:
          type: string
          title: Folder
          description: Absolute path to a photo folder
      type: object
      required:
      - folder
      title: RootRequest
    SavedSearch:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        request:
          $ref: '#/components/schemas/SearchRequest'
      type: object
      required:
      - name
      - request
      title: SavedSearch
    SearchRequest:
      properties:
        favorites_only:
          type: boolean
          title: Favorites Only
          default: false
        min_rating:
          type: integer
          maximum: 5.0
          minimum: 0.0
          title: Min Rating
          default: 0
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
          description: Natural-language description
        start_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Date
        people_ids:
          items:
            type: integer
          type: array
          title: People Ids
        people_mode:
          type: string
          enum:
          - any
          - all
          title: People Mode
          default: any
        has_location:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Location
        has_faces:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Faces
        folder:
          anyOf:
          - type: string
          - type: 'null'
          title: Folder
        camera:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera
        untagged_only:
          type: boolean
          title: Untagged Only
          default: false
        near_lat:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Near Lat
        near_lon:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Near Lon
        near_km:
          type: number
          maximum: 20000.0
          exclusiveMinimum: 0.0
          title: Near Km
          default: 1.0
        media:
          anyOf:
          - type: string
            enum:
            - image
            - video
          - type: 'null'
          title: Media
        sort:
          type: string
          enum:
          - relevance
          - quality
          - date_desc
          - date_asc
          - added_desc
          - random
          title: Sort
          default: relevance
        page:
          type: integer
          minimum: 1.0
          title: Page
          default: 1
        per_page:
          anyOf:
          - type: integer
            maximum: 1000.0
            minimum: 1.0
          - type: 'null'
          title: Per Page
        min_score:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: -1.0
          - type: 'null'
          title: Min Score
          description: Drop semantic matches below this cosine similarity
      type: object
      title: SearchRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError